tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/constants/004.html

<!doctype html>
<title>WebSockets: getting constants on prototype and object</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
var constants = ['CONNECTING', 'OPEN', 'CLOSING', 'CLOSED'];
for (var i = 0; i < constants.length; ++i) {
  test(function() {
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
    assert_equals(WebSocket.prototype[constants[i]], i);
   }, 'WebSocket.prototype.'+constants[i]);
   test(function() {
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
    assert_equals(ws[constants[i]], i);
  }, 'ws.'+constants[i]);
};
</script>
